--- a/lib/galaxy/model/orm/engine_factory.py
+++ b/lib/galaxy/model/orm/engine_factory.py
@@ -1,4 +1,5 @@
 import logging
+import threading
 import time
 
 from sqlalchemy import create_engine, event
@@ -6,8 +7,24 @@ from sqlalchemy.engine import Engine
 
 log = logging.getLogger(__name__)
 
+QUERY_COUNT_LOCAL = threading.local()
 
-def build_engine(url, engine_options, database_query_profiling_proxy=False, trace_logger=None, slow_query_log_threshold=0, thread_local_log=None):
+
+def reset_request_query_counts():
+    QUERY_COUNT_LOCAL.times = []
+
+
+def log_request_query_counts(req_id):
+    try:
+        times = QUERY_COUNT_LOCAL.times
+        if times:
+            log.info("Executed [%s] SQL requests in for web request [%s] (%f(s)ms)" % (len(times), req_id, sum(times) * 1000.))
+    except AttributeError:
+        # Didn't record anything so don't worry.
+        pass
+
+
+def build_engine(url, engine_options, database_query_profiling_proxy=False, trace_logger=None, slow_query_log_threshold=0, thread_local_log=None, log_query_counts=False):
     # Should we use the logging proxy?
     if database_query_profiling_proxy:
         import galaxy.model.orm.logging_connection_proxy as logging_connection_proxy
@@ -18,7 +35,7 @@ def build_engine(url, engine_options, database_query_profiling_proxy=False, trac
         proxy = logging_connection_proxy.TraceLoggerProxy(trace_logger)
     else:
         proxy = None
-    if slow_query_log_threshold or thread_local_log:
+    if slow_query_log_threshold or thread_local_log or log_query_counts:
         @event.listens_for(Engine, "before_execute")
         def before_execute(conn, clauseelement, multiparams, params):
             conn.info.setdefault('query_start_time', []).append(time.time())
@@ -29,6 +46,12 @@ def build_engine(url, engine_options, database_query_profiling_proxy=False, trac
             total = time.time() - conn.info['query_start_time'].pop(-1)
             if total > slow_query_log_threshold:
                 log.debug("Slow query: %f(s)\n%s\nParameters: %s" % (total, statement, parameters))
+            if log_query_counts:
+                try:
+                    QUERY_COUNT_LOCAL.times.append(total)
+                except AttributeError:
+                    # Not a web thread.
+                    pass
             if thread_local_log is not None:
                 try:
                     if thread_local_log.log: