--- a/test/unit/tools/test_execution.py
+++ b/test/unit/tools/test_execution.py
@@ -211,6 +211,9 @@ class MockTrans(object):
     def get_current_user_roles(self):
         return []
 
+    def log_event(self, *args, **kwds):
+        pass
+
 
 class MockCollectionService(object):