Work on #4959, like #6380 & #6548  but a better set of optimizations - dropped a few things that didn't pan out and added some nice newer optimizations that are better.

Below are some timings for a collection mapping job that creates 100 datasets (10 collections of 10). This branch is getting very close to resulting in a 10x speed up in creating jobs. They are a bit interconnect though still and I want tests for the permission handling changes - so I think I want #6461 in before getting too far with this. Obviously though my initial results are that #6550 is a really nice optimization in its own right and gets us half way there.

dev

galaxy.tools.execute DEBUG 2018-08-01 14:35:00,126 Executed 10 job(s) for tool create_10 request: (13434.709 ms)

dev with populate object store in job handler (PR #6550):

galaxy.tools.execute DEBUG 2018-08-01 14:37:53,418 Executed 10 job(s) for tool create_10 request: (6131.476 ms)

dev + PR#6550 + permission optimization:

galaxy.tools.execute DEBUG 2018-08-01 14:42:08,823 Executed 10 job(s) for tool create_10 request: (5202.416 ms)

dev + PR#6550 + permission optimization + add_datasets query opts:

galaxy.tools.execute DEBUG 2018-08-01 14:33:10,409 Executed 10 job(s) for tool create_10 request: (4712.302 ms)

dev + PR#6550 + permission optimization + add_datasets query opts + flush once:

galaxy.tools.execute DEBUG 2018-08-01 15:14:11,361 Executed 10 job(s) for tool create_10 request: (1549.116 ms)