@jmchilton 2018-09-07
The only optimization left in here is the one that flushes all jobs for a collection mapping step at once (36c5ba8385790f0e82870a97c8a45297702ce33b). It is a small-big change that in the tests above cut the submission runtime time of a collection mapping step to a third of what it was without the change. The runtime becomes dominated by one really large SA flush after that change (still a lot to be done after that point - but it becomes about figuring out how to get SA to write raw mappings without re-populating object attributes we don't need instead of just avoiding flushes - which has been the main technique to optimize things up until now).